Update examples to use definitions.py (part 1) (#20771)
## Summary & Motivation

After discussion with the team, we will update our examples, docs and
Dagster University courses to recommend initializing the Definitions
object in `definitions.py` instead of `__init__.py`.

This PR is a first effort to update our examples. 

Examples that are missing from this PR and should be updated in another
one:
- assets_*
- experimental
- project_*
- tutorial_*

Examples that we should be updated when we update the docs:
- doc_snippets


## How I Tested These Changes

Each example was tested locally + BK
