+++ examples/quickstart_etl/README.md
@@ -157,7 +157,7 @@ Note: You'll find a `path` metadata attached to every asset. This is because ass
 
 Finally, let's refresh our plots every day so we can monitor popular topics over time. To do so, we can use [schedules](https://docs.dagster.io/concepts/partitions-schedules-sensors/schedules#schedules).
 
-We've defined a daily schedule and job in [`quickstart_etl/__init__.py`](./quickstart_etl/__init__.py) for all assets that are defined in the [`quickstart_etl/assets/`](./quickstart_etl/assets) module.
+We've defined a daily schedule and job in [`quickstart_etl/definitions.py`](./quickstart_etl/definitions.py) for all assets that are defined in the [`quickstart_etl/assets/`](./quickstart_etl/assets) module.
 
 Now, let's turn on the daily schedule within Dagster.
 
@@ -168,7 +168,7 @@ Now, let's turn on the daily schedule within Dagster.
     <img height="500" src="../../docs/next/public/images/quickstarts/basic/step-3-1-schedule-off.png" />
 </p>
 
-You can now turn on the schedule switch to set up the daily job we defined in [quickstart_etl/**init**.py](./quickstart_etl/__init__.py).
+You can now turn on the schedule switch to set up the daily job we defined in [quickstart_etl/definitions.py](./quickstart_etl/definitions.py).
 
 <p align="center">
     <img height="500" src="../../docs/next/public/images/quickstarts/basic/step-3-2-schedule-on.png" />
