+++ examples/quickstart_etl/quickstart_etl/definitions.py
@@ -1,0 +1,16 @@
+from dagster import (
+    Definitions,
+    ScheduleDefinition,
+    define_asset_job,
+    load_assets_from_package_module,
+)
+
+from . import assets
+
+daily_refresh_schedule = ScheduleDefinition(
+    job=define_asset_job(name="all_assets_job"), cron_schedule="0 0 * * *"
+)
+
+defs = Definitions(
+    assets=load_assets_from_package_module(assets), schedules=[daily_refresh_schedule]
+)
