+++ examples/quickstart_gcp/README.md
@@ -55,7 +55,7 @@ To connect to GCP, you'll need to set up your credentials in Dagster.
 
 Dagster allows using environment variables to handle sensitive information. You can define various configuration options and access environment variables through them. This also allows you to parameterize your pipeline without modifying code.
 
-In this example, we use [BigQueryPandasIOManager](https://docs.dagster.io/_apidocs/libraries/dagster-gcp-pandas#dagster_gcp_pandas.BigQueryPandasIOManager) to write outputs to BigQuery and read inputs from it. The configurations of the BigQuery connection are defined [in `quickstart_gcp/__init__.py`](./quickstart_snowflake/__init__.py), which requires the following environment variables:
+In this example, we use [BigQueryPandasIOManager](https://docs.dagster.io/_apidocs/libraries/dagster-gcp-pandas#dagster_gcp_pandas.BigQueryPandasIOManager) to write outputs to BigQuery and read inputs from it. The configurations of the BigQuery connection are defined [in `quickstart_gcp/definitions.py`](./quickstart_snowflake/definitions.py), which requires the following environment variables:
 
 - `BIGQUERY_PROJECT_ID`
